requirements.txt
fastapi
uvicorn
python-multipart
jinja2
sentence-transformers
chromadb
PyPDF2
google-generativeai
mistralai
python-dotenv